tagscope/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
duckdb==1.5.2
pandas==2.2.0
pyyaml==6.0.1
